requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.13.0
appwrite==16.0.0
arabic-reshaper==3.0.1
asn1crypto==1.5.1
Authlib==1.7.0
bidict==0.23.1
blinker==1.9.0
cerebras_cloud_sdk==1.67.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==47.0.0
cssselect2==0.9.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
eventlet==0.41.0
gunicorn==23.0.0
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Mailman==1.1.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
fonttools==4.62.1
fpdf2==2.8.7
freetype-py==2.5.1
greenlet==3.3.2
groq==1.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
joserfc==1.6.4
lxml==6.1.0
Mako==1.3.10
markdown2==2.5.5
MarkupSafe==3.0.3
openai==2.33.0
oscrypto==1.3.0
pillow==12.2.0
psycopg==3.3.3
psycopg-c==3.3.3
pycairo==1.29.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyHanko==0.35.0
pyhanko-certvalidator==0.31.0
PyJWT==2.12.0
pypdf==6.10.2
python-bidi==0.6.7
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
PyYAML==6.0.3
reportlab==4.5.0
requests==2.32.5
rlPyCairo==0.4.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
svglib==1.6.0
tinycss2==1.5.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.1.0
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.6
wsproto==1.3.2
xhtml2pdf==0.2.17
